+++ app/src/main/java/utils/Saloon.java
@@ -9,7 +9,7 @@ import java.util.Map;
 
 public class Saloon implements Serializable {
 
-    private long saloonPoint;
+    private int saloonPoint;
     private String saloonLocation;
     private int saloonRating;
     private String saloonName;
@@ -57,11 +57,11 @@ public class Saloon implements Serializable {
 
     }
 
-    public long getSaloonPoint() {
+    public int getSaloonPoint() {
         return saloonPoint;
     }
 
-    public void setSaloonPoint(long saloonPoint) {
+    public void setSaloonPoint(int saloonPoint) {
         this.saloonPoint = saloonPoint;
     }
